Move helpers and requests out of components folder
