+++ src/public/components/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Routes from './Routes';
 import Navigation from './Navigation';
-import { getUserData } from './requests/users';
+import { getUserData } from '../requests/users';
 import { BrowserRouter as Router, Route, Link, Redirect } from 'react-router-dom';
 
 export default class App extends React.Component {
