+++ src/public/helpers/tableHelpers.js
